requirements.txt
python-dotenv~=1.0.1
fastapi~=0.111.0
SQLAlchemy~=2.0.30
pydantic~=2.7.4
starlette~=0.37.2